Task 2.17 done with dialog

+++ osa2/puhelinluettelo/src/App.js
@@ -49,12 +49,13 @@ const App = () => {
   }
 
   const deleteHandler = (person) => {
-    console.log(person)
-    personService
-      .deletePers(person)
-      .then(response => {
-        setPersons(persons.filter(pers => pers.id !== person.id))
-      })
+    if (window.confirm(`Delete ${person.name}?`)) {
+      personService
+        .deletePers(person)
+        .then(response => {
+          setPersons(persons.filter(pers => pers.id !== person.id))
+        })
+    }
   }
 
   return (
